relocated stepmatcher classes
