+++ org.agileware.natural.common/META-INF/MANIFEST.MF
@@ -5,15 +5,8 @@ Bundle-SymbolicName: org.agileware.natural.common
 Bundle-Version: 1.3.0.qualifier
 Bundle-Activator: org.agileware.natural.common.Activator
 Require-Bundle: org.eclipse.core.runtime;bundle-version="3.17.100";visibility:=reexport,
- org.eclipse.jface.text;bundle-version="3.16.200";visibility:=reexport,
- org.eclipse.ui;bundle-version="3.116.0",
- org.eclipse.ui.ide;bundle-version="3.17.0",
- org.eclipse.jdt.ui;bundle-version="3.21.0",
- org.eclipse.jdt.core;bundle-version="3.21.0",
- org.eclipse.core.resources;bundle-version="3.13.700",
  com.google.guava;bundle-version="27.1.0",
- com.google.inject;bundle-version="3.0.0",
- org.eclipse.emf.ecore
+ com.google.inject;bundle-version="3.0.0"
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-Vendor: Roberto Lo Giacco
