+++ org.agileware.natural.cucumber.ui/src/org/agileware/natural/cucumber/ui/CucumberHyperlinkHelper.java
@@ -5,9 +5,9 @@ import java.util.Arrays;
 import java.util.Collection;
 import java.util.List;
 
-import org.agileware.natural.common.JavaAnnotationMatcher;
-import org.agileware.natural.common.JavaHyperlink;
 import org.agileware.natural.cucumber.cucumber.Step;
+import org.agileware.natural.stepmatcher.JavaAnnotationMatcher;
+import org.agileware.natural.stepmatcher.JavaHyperlink;
 import org.eclipse.emf.ecore.EObject;
 import org.eclipse.jdt.core.IMethod;
 import org.eclipse.jface.text.Region;
