+++ org.agileware.natural.cucumber/META-INF/MANIFEST.MF
@@ -7,6 +7,7 @@ Bundle-Version: 1.0.0.qualifier
 Bundle-SymbolicName: org.agileware.natural.cucumber; singleton:=true
 Bundle-ActivationPolicy: lazy
 Require-Bundle: org.agileware.natural.common;visibility:=reexport,
+ org.agileware.natural.stepmatcher;visibility:=reexport,
  org.eclipse.xtext,
  org.eclipse.xtext.xbase,
  org.eclipse.equinox.common;bundle-version="3.5.0",
