+++ org.agileware.natural.cucumber/src/org/agileware/natural/cucumber/CucumberRuntimeModule.java
@@ -3,7 +3,7 @@
  */
 package org.agileware.natural.cucumber;
 
-import org.agileware.natural.common.AbstractAnnotationDescriptor;
+import org.agileware.natural.stepmatcher.AbstractAnnotationDescriptor;
 
 /**
  * Use this class to register components to be used at runtime / without the Equinox extension registry.
