+++ org.agileware.natural.stepmatcher/META-INF/MANIFEST.MF
@@ -4,7 +4,22 @@ Bundle-Name: Stepmatcher
 Bundle-SymbolicName: org.agileware.natural.stepmatcher
 Bundle-Version: 1.0.0.qualifier
 Bundle-Activator: org.agileware.natural.stepmatcher.Activator
-Require-Bundle: org.eclipse.core.runtime
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.17.100";visibility:=reexport,
+ org.eclipse.jface.text;bundle-version="3.16.200";visibility:=reexport,
+ org.eclipse.ui;bundle-version="3.116.0",
+ org.eclipse.ui.ide;bundle-version="3.17.0",
+ org.eclipse.jdt.ui;bundle-version="3.21.0",
+ org.eclipse.jdt.core;bundle-version="3.21.0",
+ org.eclipse.core.resources;bundle-version="3.13.700",
+ com.google.guava;bundle-version="27.1.0",
+ com.google.inject;bundle-version="3.0.0",
+ org.eclipse.emf.ecore
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Automatic-Module-Name: org.agileware.natural.stepmatcher
 Bundle-ActivationPolicy: lazy
+Export-Package: org.agileware.natural.stepmatcher;
+  uses:="org.osgi.framework,
+   org.eclipse.jdt.core,
+   org.eclipse.jface.text,
+   org.eclipse.jface.text.hyperlink,
+   org.eclipse.jdt.core.search"
