+++ org.agileware.natural.stepmatcher/src/org/agileware/natural/stepmatcher/AbstractAnnotationDescriptor.java
@@ -1,4 +1,4 @@
-package org.agileware.natural.common;
+package org.agileware.natural.stepmatcher;
 
 import org.eclipse.jdt.core.IJavaElement;
 import org.eclipse.jdt.core.IType;
