+++ org.agileware.natural.stepmatcher/src/org/agileware/natural/stepmatcher/JavaAnnotationMatcher.java
@@ -1,4 +1,4 @@
-package org.agileware.natural.common;
+package org.agileware.natural.stepmatcher;
 
 import java.util.ArrayList;
 import java.util.Collection;
