+++ org.agileware.natural.stepmatcher/src/org/agileware/natural/stepmatcher/JavaElementChangeReporter.java
@@ -1,4 +1,4 @@
-package org.agileware.natural.common;
+package org.agileware.natural.stepmatcher;
 
 import java.util.List;
